a breaking change to only have a single, default export (#31)
there were only 3 cases in our codebase where we were using a named import from this library, and they have been updated to not use it anymore.

the reasoning is that mixing named and default exports causes problems in certain module formats and usages, so if we remove one or the other, it works and we can then use the format:system bundles for this library in our app. that then allows us to start using that format for our microservices as well.