the real change here is the new exported const __esModule (#33)
why? well, a long story short is that other apps NEED that value to know that single-spa-canopy has a default export. wihtout it, they wrap single-spa-canopy in another object that looks like this: {default: [module]}. so you have to do a .default.default. ugh.

normally these things are handled by babel and/or bundlers (webpack, rollup). however, in our case, for some reason rollup has decided NOT to add that when you only have a default export. so we have to add it manually.